refactor: userDetails와 user 엔티티 분리
#60
